+++ src/main/java/tonius/simplyjetpacks/capability/EnergyConversionStorage.java
@@ -1,6 +1,5 @@
 package tonius.simplyjetpacks.capability;
 
-import cofh.redstoneflux.api.IEnergyContainerItem;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.energy.IEnergyStorage;